web-app/pyproject.toml
[tool.poetry]
name = "web-app"
version = "0.1.0"
description = ""
authors = ["Kevin Chen <kevc528@seas.upenn.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
SQLAlchemy = "^1.4.9"
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.1"
psycopg2 = "^2.8.6"
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.17.0"
alembic = "^1.5.8"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
